refactor: activity 명 변경
