+++ app/src/main/java/woowacourse/omok/presentation/GameActivity.kt
@@ -1,4 +1,4 @@
-package woowacourse.omok
+package woowacourse.omok.presentation
 
 import android.os.Bundle
 import android.widget.ImageView
@@ -11,6 +11,7 @@ import androidx.appcompat.app.AppCompatActivity
 import androidx.core.view.ViewCompat
 import androidx.core.view.WindowInsetsCompat
 import androidx.core.view.children
+import woowacourse.omok.R
 import woowacourse.omok.data.db.BoardDao
 import woowacourse.omok.data.db.DbHelper
 import woowacourse.omok.domain.OmokAdapter
@@ -23,7 +24,7 @@ import woowacourse.omok.domain.PutStoneResult.NextTurn
 import woowacourse.omok.domain.Stone
 import woowacourse.omok.domain.StoneState
 
-class MainActivity : AppCompatActivity() {
+class GameActivity : AppCompatActivity() {
     private lateinit var omokGame: OmokGame
     private lateinit var board: TableLayout
     private lateinit var dbHelper: DbHelper
@@ -33,7 +34,7 @@ class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         enableEdgeToEdge()
-        setContentView(R.layout.activity_main)
+        setContentView(R.layout.activity_game)
         ViewCompat.setOnApplyWindowInsetsListener(findViewById(R.id.main)) { v, insets ->
             val systemBars = insets.getInsets(WindowInsetsCompat.Type.systemBars())
             v.setPadding(systemBars.left, systemBars.top, systemBars.right, systemBars.bottom)
